refactor: src directorie updated build
