+++ src/check.ts
